CoffeeShopSituation 구현 에서 본인 주문 맞는지 확인 로직 추가

+++ src/exercise/chapter_29/Cashier.java
@@ -50,7 +50,7 @@ public class Cashier {
     }
 
     //3. 돈을 더한다.
-    public void addAmount(int money){
+    public void addAmount(long money){
         //내 돈에서 더하는 행위
         this.salesAmount += money;
     }
